v1.067
v1.067

(Experimental Update)

Your settings save file will now crash your game. To fix this go to the top of the op objects create event and set the load variable to false. Then launch the game, start a new save file, go into the settings menu in the dark world and modify the volume and then exit out of the settings menu. This should create a new settings save file. Close the game and set the load variable back to ture.